update reward function
Just give a positive reward if it gets closer, don't base reward on distance covered.
Once the robot got really close to the target it could only cover a very small distance per time step,
whereas in the start it could cover a lot per time step. So to make the rewards have the same order of magnitude across time steps we just go with +1 or -1.